Conflicts:
	ShouKeBao/Images.xcassets/jianbian.imageset/Contents.json
	ShouKeBao/Images.xcassets/收客宝切图/banjiubing2.imageset/Contents.json
	ShouKeBao/Images.xcassets/收客宝切图/tianjia.imageset/Contents.json
	ShouKeBao/ShouKebao/ShouKeBao.m
	ShouKeBao/ShouKebao/ShouKeBao.xib

+++ ShouKeBao/Tool/Animations/WMAnimations.m
@@ -62,7 +62,7 @@ theAnimation.fromValue=[NSValue valueWithCGPoint:*fromPoint];
 
 }
 
-+ (void)WMAnimationMakeBoarderWithLayer:(CALayer *)layer andBorderColor:(UIColor *)color andBorderWidth:(int)borderWid andNeedShadow:(BOOL)needShow
++ (void)WMAnimationMakeBoarderWithLayer:(CALayer *)layer andBorderColor:(UIColor *)color andBorderWidth:(float)borderWid andNeedShadow:(BOOL)needShow
 {
     layer.borderColor = color.CGColor;
     layer.borderWidth = borderWid;
